LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/snap-into-covering-area-expected.txt

scrollerButton

FAIL snap-into-covering-area assert_equals: scroller is snapped to the bottom of the larger snap area, not the top expected 4367 but got 0
